==> notebooks/modulo_classificador_nlu/04_active_learning_extratificacao.ipynb <==
### Visando melhorar a eficiência do treinamento e anotação para o tempo hábil do projeto, decidimos focar em uma amostra estratificada do dataset completo. Esta abordagem nos permite capturar a diversidade de perguntas e tópicos presentes no dataset original, garantindo que o modelo seja exposto a uma variedade representativa de exemplos sem a necessidade de processar o conjunto completo.

Usando as mesmas proporções da análise exploratória, aplicamos ao nosso novo alvo de 1.000:
    - Ciências da Natureza: 1.000 * 0.279 = 279 exemplos
    - Ciências Humanas: 1.000 * 0.265 = 265 exemplos
    - Linguagens e Códigos: 1.000 * 0.236 = 236 exemplos
    - Matemática: 1.000 * 0.221 = 221 exemplos

    Total: 1.001 exemplos.

import pandas as pd

# Carregar o dataset completo
df = pd.read_csv('/Users/giossaurus/Developer/leia_tcc/data/interim/simulados_enem_completo.csv')

# Definir o número de amostras para o universo de trabalho de 1.000
working_set_samples = {
    'Ciências da Natureza': 279,
    'Ciências Humanas': 265,
    'Linguagens e Códigos': 236,
    'Matemática': 221
}

# Criar a amostra estratificada
df.rename(columns={'dataset': 'disciplina'}, inplace=True)

sampled_dfs = []
for subject, n_samples in working_set_samples.items():
    subject_df = df[df['disciplina'] == subject]
    sampled_dfs.append(subject_df.sample(n=n_samples, random_state=42))

# Concatenar e embaralhar o dataframe final
working_set_df = pd.concat(sampled_dfs)
working_set_df = working_set_df.sample(frac=1, random_state=42).reset_index(drop=True)

# Salvar o arquivo do universo de trabalho
working_set_df.to_csv('/Users/giossaurus/Developer/leia_tcc/data/processed/amostra_trabalho_1000.csv', index=False)

print("Amostra de trabalho com 1.001 exemplos criada com sucesso!")
print(f"Arquivo salvo com sucesso")
